requirements.txt
graphql-core>=3.2
mitmproxy>=10.2
